requirements.txt
# Airflow
apache-airflow==2.8.1
apache-airflow-providers-google
apache-airflow-providers-apache-spark

# Data handling
pandas
pyarrow
requests
beautifulsoup4
selenium

# GCP
google-cloud-storage
google-cloud-bigquery-storage

# Web parsing
lxml
html5lib